refactor: code adjustments based on gh comment
