+++ src/components/language-buttons/styles.ts
@@ -1,0 +1,12 @@
+import { tss } from '@/styles/theme';
+
+export const useStyles = tss.create(() => {
+  return {
+    languageButtons: {
+      marginLeft: 'auto',
+    },
+    button: {
+      margin: '0 5px',
+    },
+  };
+});
