+++ src/layouts/header/header.tsx
@@ -1,9 +1,6 @@
-import { useState, useEffect } from 'react';
-import { useTranslation } from 'react-i18next';
+import { Box, Typography } from '@mui/material';
 
-import { changeLanguage } from 'i18next';
-
-import { Box, Button, Typography } from '@mui/material';
+import { LanguageButtons } from '@/components/language-buttons';
 
 import { useStyles } from './styles';
 
@@ -13,49 +10,11 @@ interface IHeaderProps {
 
 export function Header({ title }: IHeaderProps): React.ReactNode {
   const { classes } = useStyles();
-  const { i18n } = useTranslation();
-  const [selectedLanguage, setSelectedLanguage] = useState<string>(
-    i18n.language,
-  );
-
-  useEffect(() => {
-    setSelectedLanguage(i18n.language);
-  }, [i18n.language]);
-
-  function handleChangeLanguage(lang: string): void {
-    void changeLanguage(lang);
-    setSelectedLanguage(lang);
-  }
-
-  function getVariant(lang: string): 'contained' | 'outlined' {
-    return selectedLanguage === lang ? 'contained' : 'outlined';
-  }
 
   return (
     <Box className={classes.header}>
       <Typography className={classes.title}>{title}</Typography>
-      <Box className={classes.languageButtons}>
-        <Button
-          size="small"
-          className={classes.button}
-          onClick={(): void => {
-            return handleChangeLanguage('en');
-          }}
-          variant={getVariant('en')}
-        >
-          EN
-        </Button>
-        <Button
-          size="small"
-          className={classes.button}
-          onClick={(): void => {
-            return handleChangeLanguage('de');
-          }}
-          variant={getVariant('de')}
-        >
-          DE
-        </Button>
-      </Box>
+      <LanguageButtons />
     </Box>
   );
 }
