+++ src/pages/order-page/order-page.tsx
@@ -4,12 +4,12 @@ import { Box } from '@mui/material';
 
 import { Header } from '@/layouts/header';
 
-import { useStyles } from './styles';
+import { useStyles } from '../styles';
 
 /* eslint-disable import/no-default-export */
 export default function OrderPage(): React.ReactNode {
-  const { classes } = useStyles();
   const { t } = useTranslation();
+  const { classes } = useStyles();
 
   return (
     <Box className={classes.wrapper}>
